+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - :sparkles: Config file `site/config/panelBar.yml`
 - :sparkles: Modal component to be used by elements
 - :sparkles: Content component for easily styled text
+- :sparkles: New element: `navigation`
 - :sparkles: Login icon when user is not signed-in
 - :balloon: Dropdown pattern: additional CSS classes per dropdown item
 - :balloon: Added `about` element as sample for modal and content component